refactor(core): extract theme configuration into separate file
Separate theme configuration from theme context to improve code organization and maintainability. Move theme type definitions and default theme objects to dedicated theme-config.ts file.
